+++ packages/orq-rc/src/models/errors/index.ts
@@ -51,6 +51,7 @@ export * from "./getpromptversion.js";
 export * from "./honoapierror.js";
 export * from "./httpclienterrors.js";
 export * from "./invokeeval.js";
+export * from "./listagenttasks.js";
 export * from "./orqerror.js";
 export * from "./responsevalidationerror.js";
 export * from "./retrievecontact.js";