+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -12,7 +12,7 @@ import { SDKValidationError } from "../errors/sdkvalidationerror.js";
 /**
  * Filter datasources by status.
  */
-export type Status = Array<string> | string;
+export type QueryParamStatus = Array<string> | string;
 
 export type ListDatasourcesRequest = {
   /**
@@ -109,43 +109,50 @@ export type ListDatasourcesResponseBody = {
 };
 
 /** @internal */
-export const Status$inboundSchema: z.ZodType<Status, z.ZodTypeDef, unknown> = z
-  .union([z.array(z.string()), z.string()]);
-
-/** @internal */
-export type Status$Outbound = Array<string> | string;
-
-/** @internal */
-export const Status$outboundSchema: z.ZodType<
-  Status$Outbound,
+export const QueryParamStatus$inboundSchema: z.ZodType<
+  QueryParamStatus,
   z.ZodTypeDef,
-  Status
+  unknown
+> = z.union([z.array(z.string()), z.string()]);
+
+/** @internal */
+export type QueryParamStatus$Outbound = Array<string> | string;
+
+/** @internal */
+export const QueryParamStatus$outboundSchema: z.ZodType<
+  QueryParamStatus$Outbound,
+  z.ZodTypeDef,
+  QueryParamStatus
 > = z.union([z.array(z.string()), z.string()]);
 
 /**
  * @internal
  * @deprecated This namespace will be removed in future versions. Use schemas and types that are exported directly from this module.
  */
-export namespace Status$ {
-  /** @deprecated use `Status$inboundSchema` instead. */
-  export const inboundSchema = Status$inboundSchema;
-  /** @deprecated use `Status$outboundSchema` instead. */
-  export const outboundSchema = Status$outboundSchema;
-  /** @deprecated use `Status$Outbound` instead. */
-  export type Outbound = Status$Outbound;
+export namespace QueryParamStatus$ {
+  /** @deprecated use `QueryParamStatus$inboundSchema` instead. */
+  export const inboundSchema = QueryParamStatus$inboundSchema;
+  /** @deprecated use `QueryParamStatus$outboundSchema` instead. */
+  export const outboundSchema = QueryParamStatus$outboundSchema;
+  /** @deprecated use `QueryParamStatus$Outbound` instead. */
+  export type Outbound = QueryParamStatus$Outbound;
 }
 
-export function statusToJSON(status: Status): string {
-  return JSON.stringify(Status$outboundSchema.parse(status));
+export function queryParamStatusToJSON(
+  queryParamStatus: QueryParamStatus,
+): string {
+  return JSON.stringify(
+    QueryParamStatus$outboundSchema.parse(queryParamStatus),
+  );
 }
 
-export function statusFromJSON(
+export function queryParamStatusFromJSON(
   jsonString: string,
-): SafeParseResult<Status, SDKValidationError> {
+): SafeParseResult<QueryParamStatus, SDKValidationError> {
   return safeParse(
     jsonString,
-    (x) => Status$inboundSchema.parse(JSON.parse(x)),
-    `Failed to parse 'Status' from JSON`,
+    (x) => QueryParamStatus$inboundSchema.parse(JSON.parse(x)),
+    `Failed to parse 'QueryParamStatus' from JSON`,
   );
 }
 
@@ -278,7 +285,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K6MDVPQ4F667RAQF10WFDATE"),
+  _id: z.string().default("01K6N5KY0ZBGY6P4YQE7D30BJ4"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,
@@ -322,7 +329,7 @@ export const ListDatasourcesData$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListDatasourcesData
 > = z.object({
-  id: z.string().default("01K6MDVPQ4F667RAQF10WFDATE"),
+  id: z.string().default("01K6N5KY0ZBGY6P4YQE7D30BJ4"),
   displayName: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$outboundSchema,