+++ packages/orq-rc/RELEASES.md
@@ -4069,3 +4069,13 @@ Based on:
 - [typescript v3.14.0-rc.6] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.6 - packages/orq-rc
+
+## 2025-10-03 13:55:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.634.1 (2.721.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.7] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.7 - packages/orq-rc